online_catalog/requirements.txt
Django==3.2
django-cors-headers==3.7.0
django-debug-toolbar==2.2
django-filter==2.4.0
django-import-export==2.5.0
django-utils-six==2.0
djangorestframework==3.12.3
djangorestframework-simplejwt==4.6.0
httplib2==0.19.0
isort==5.8.0
pep8==1.7.1
Pillow==8.0.0
pluggy==0.13.1
py==1.8.1
pytest==5.3.5
pytest-django==3.8.0
python-dotenv==0.15.0
requests==2.22.0
rsa==4.7.2
sqlparse==0.3.1